Make admin users room owners even without token
It may be important that the 'focus' Jicofo user can be an owner in
addition to people with moderator tokens. Without that, if the
moderator leaves the room (eg looses connection) then another user
will be made the new moderator, even if their token doesn't say they
should be. This change allows the focus user to be the owner so that
won't happen, so long as they're in the 'admins' setting in the
Prosody configuration.

Source: https://github.com/nvonahsen/jitsi-token-moderation-plugin/commit/a5ebdfaa38a6adde6bceba62cfbc5b1693e480b9